Refactored into React
