+++ client/App.jsx
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class App extends Component {
+	render() {
+		return (
+			<div>
+				If you can see me, that means REACT works!
+			</div>
+		)
+	}
+}
+
+export default App;
