+++ client/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { render } from 'react-dom'
+import App from './App';
+
+render(
+	<App />,
+	document.getElementById('root')
+)
