+++ package.json
@@ -50,18 +50,29 @@
 		"compile": "tsc -p ./",
 		"watch": "tsc -watch -p ./",
 		"postinstall": "node ./node_modules/vscode/bin/install",
-		"test": "npm run compile && node ./node_modules/vscode/bin/test"
+		"test": "npm run compile && node ./node_modules/vscode/bin/test",
+		"build": "webpack --watch"
 	},
 	"devDependencies": {
 		"typescript": "^3.3.1",
 		"vscode": "^1.1.28",
 		"tslint": "^5.12.1",
 		"@types/node": "^10.12.21",
-		"@types/mocha": "^2.2.42"
+		"@types/mocha": "^2.2.42",
+		"webpack": "^4.29.2",
+		"webpack-cli": "^3.2.3",
+		"@babel/core": "^7.2.2",
+		"@babel/preset-env": "^7.3.1",
+		"@babel/preset-react": "^7.0.0",
+		"babel-loader": "^8.0.5",
+		"css-loader": "^2.1.0",
+		"style-loader": "^0.23.1"
 	},
 	"dependencies": {
 		"chrome-launcher": "^0.10.5",
 		"electron-devtools-installer": "^2.2.4",
-		"puppeteer": "^1.12.2"
+		"puppeteer": "^1.12.2",
+		"react": "^16.5.2",
+		"react-dom": "^16.5.2"
 	}
 }
