+++ webpack.config.js
@@ -1,0 +1,36 @@
+const path = require('path')
+
+module.exports = {
+  mode: 'development',
+  entry: './client/index.js',
+  output: {
+    path: path.resolve(__dirname, 'out', 'build'),
+    filename: 'bundle.js'
+  },
+  module: {
+    rules: [
+      {
+        test: /.(js|jsx)$/,
+        exclude: /node_modules/,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets:
+              [
+                '@babel/preset-react',
+                '@babel/preset-env'
+              ]
+          }
+        }
+      },
+      {
+        test: /.css$/,
+        exclude: /node_modules/,
+        use: ['style-loader', 'css-loader']
+      }
+    ]
+  },
+  resolve: {
+    extensions: ['*', '.js', '.jsx']
+  }
+}
